플레이어 오브젝트와 충돌 수정, ui 플레이어와 연결

+++ SunHaven/SunHaven/Class/UI/UI.h
@@ -32,7 +32,7 @@ public:
 	void update(void);
 	void render(void);
 	
-	void renderBar(int x, int y, int width, int height, int alph, string Color);
+	void renderBar(int x, int y, float width, float height, int alph, string Color);
 	void showBasicUI();
 	void updateUIWindow();
 	void showUIWindow();
